Fix signed-shift UB in the MP4 box readers; add a sanitizer build
Chasing a server that dies after ~6 hours. A soak against DASH-IF's
livesim2 (player attached, 8+ minutes) showed flat RSS, flat fd count and
clean playback, and forcing the seen-set FIFO eviction -- the one path
that normally waits ~4.5 hours to run -- changed nothing. Memory staying
flat rules out exhaustion and points at a bad pointer instead, so the MP4
parsers that chew on origin-supplied bytes got fuzzed under
ASan + UBSan: every truncation length, single-byte corruption across the
box headers, and oversized size fields, on a real init and media segment.

ASan found no out-of-bounds access -- the parsers are bounds-safe. UBSan
found this:

  core/src/cenc.c:10: left shift of 255 by 24 places cannot be
  represented in type 'int'

cenc.c's read_u32 and cdm.c's PSSH kid_count both shift a uint8_t left by
24 without widening first, so the operand is promoted to int and any box
whose size begins with a byte >= 0x80 is a signed overflow. read_u32 is
the size field of every box of every segment the engine decrypts. The
behaviour is undefined, and an optimising build -- which is what runs in
production, unlike the -O1 build under test -- may legally do something
other than the wraparound that happens to work here.

This is not proven to be the crash: no reproduction, and the UB may well
have been benign under this compiler. It is a real defect on the hot path
either way. To settle it rather than keep guessing, -DRS_SANITIZE=ON now
builds the server with ASan + UBSan so the next occurrence aborts at the
faulting instruction with a stack instead of somewhere unrelated later.

+++ core/src/cdm.c
@@ -89,9 +89,14 @@ static void classify_pssh(rs_drm_challenge *ch, const char *b64) {
                 }
                 
                 if (bytes[8] == 1 && len >= 32) {
-                    int kid_count = (bytes[28] << 24) | (bytes[29] << 16) | (bytes[30] << 8) | bytes[31];
+                    // Widen before shifting: `bytes[28] << 24` on a byte >= 0x80
+                    // is a signed overflow, and the count is attacker-supplied.
+                    // The loop bounds below already contain it; this keeps the
+                    // arithmetic itself defined.
+                    uint32_t kid_count = ((uint32_t)bytes[28] << 24) | ((uint32_t)bytes[29] << 16)
+                                       | ((uint32_t)bytes[30] << 8) | (uint32_t)bytes[31];
                     size_t offset = 32;
-                    for (int i = 0; i < kid_count && i < 64 && offset + 16 <= len; i++) {
+                    for (uint32_t i = 0; i < kid_count && i < 64 && offset + 16 <= len; i++) {
                         char *kid_hex = bytes_to_hex(bytes + offset, 16);
                         if (kid_hex) {
                             add_kid(ch, kid_hex);
